chore: refactor Image Comparator to Image Comparer a bit more..
