+++ tools/image-comparer/src/ImageComparer.vue
